requirements.txt
python-telegram-bot[job-queue]==20.0
psycopg2-binary
requests>=2.31
yfinance==0.2.54
numpy>=1.23.2,<2
pandas==2.2.3
matplotlib==3.8.3
plotly==6.0.0
flask==3.1.0
fastapi==0.115.11
uvicorn==0.34.0
beautifulsoup4==4.13.3
aiogram==3.3.0
python-dotenv